refactor: rename file and moving struct
