+++ src/pkg/structs/structs_products.go
@@ -1,0 +1,10 @@
+package structs_products
+
+type Product struct {
+	Id       int64  `json:"id"`
+	Nome     string `json:"nome"`
+	Gtin     string `json:"gtin"`
+	Inclusao string `json:"inclusao"`
+}
+
+type Products []Product
